==> 00_Preprocess_PaperBestFOS.ipynb <==
import pandas as pd
import gc

FOS_DATA = pd.read_csv("./Processed_data/MAG_PaperJournalFiledDataMergedL2.tsv", sep="\t", index_col = 0)

gc.collect()

FOS_DATA

FOS_DATA_maxscore = FOS_DATA.sort_values('Score', ascending=False).drop_duplicates(['PaperId'])

FOS_DATA_maxscore.to_csv("./Processed_data/MAG_PaperFieldL2_MAX.txt", sep = "\t", index = None)
