==> 00_Preprocess_PaperFieldAssign.ipynb <==
import pandas as pd
import csv
from collections import defaultdict
from tqdm import tqdm_notebook
import gc
import numpy as np

FOS_datafile = "./MAG_dataset/advanced/FieldsOfStudy.txt"
df_fosinfo = pd.read_csv(FOS_datafile, sep="\t", header=None, usecols=[0, 3, 5, 6], quoting=csv.QUOTE_NONE)
df_fosinfo.columns = ["FieldofStudyId", "FosName", "FosLevel", "PaperCount"]

PaperData = pd.read_csv("./MAG_dataset/mag/Papers.txt", sep ="\t", error_bad_lines=False, usecols =[0, 3, 7], header=None, quoting=csv.QUOTE_NONE)
PaperData.columns = ["PaperId", "DocType", "Year"]

PaperData = PaperData[PaperData["DocType"] == "Journal"]

PaperFieldData = pd.read_csv("./MAG_dataset/advanced/PaperFieldsOfStudy.txt", sep ="\t", error_bad_lines=False, header=None, quoting=csv.QUOTE_NONE)
PaperFieldData.columns = ["PaperId", "FieldofStudyId", "Score"]

PaperFieldData = PaperFieldData.merge(df_fosinfo, on="FieldofStudyId") 

PaperFieldDataMerged = PaperData.merge(PaperFieldData, on="PaperId")

PaperFieldDataMerged = PaperFieldDataMerged.dropna(subset=["FieldofStudyId"])

PaperFieldDataMerged.drop("PaperCount", axis = 1)

PaperFieldDataMerged.to_csv("./Processed_data/MAG_PaperJournalFiledDataMerged.tsv", sep = "\t")

for i in range(0, 6):
     PaperFieldDataMerged[PaperFieldDataMerged["FosLevel"] == i].to_csv("./Processed_data/MAG_PaperJournalFiledDataMergedL" + str(i) +".tsv", sep = "\t")
