==> 00_Preprocess_SelectCittaions.ipynb <==
import pandas as pd
import csv

PaperL0FieldData = pd.read_csv("./Processed_data/MAG_PaperJournalFiledDataMergedL0.tsv", sep ="\t", index_col = 0, quoting=csv.QUOTE_NONE)

PaperL0FieldData

FOS_datafile = "./MAG_dataset/advanced/FieldsOfStudy.txt"
df_fosinfo = pd.read_csv(FOS_datafile, sep="\t", header=None, usecols=[0, 3, 5, 6], quoting=csv.QUOTE_NONE)
df_fosinfo.columns = ["FieldofStudyId", "FosName", "FosLevel", "PaperCount"]


L0_FOSID_List = df_fosinfo[df_fosinfo["FosLevel"] == 0]["FieldofStudyId"].tolist()

FOSID_NAME_dict = pd.Series(df_fosinfo["FosName"].values, index=df_fosinfo["FieldofStudyId"]).to_dict()

PaperReferences = pd.read_csv("./MAG_dataset/mag/PaperReferences.txt", sep = "\t", header = None, quoting=csv.QUOTE_NONE)

PaperReferences.merge(PaperL0FieldData[["PaperId", "FieldofStudyId"]], left_on=0, right_on="PaperId").merge(PaperL0FieldData[["PaperId", "FieldofStudyId"]], left_on=1, right_on="PaperId").drop(0, axis=1).drop(1, axis=1)

PaperReferencesFields = PaperReferences.merge(PaperL0FieldData[["PaperId", "FieldofStudyId"]], left_on=0, right_on="PaperId").merge(PaperL0FieldData[["PaperId", "FieldofStudyId"]], left_on=1, right_on="PaperId").drop(0, axis=1).drop(1, axis=1)

PaperReferencesFields.columns = ["CitingId", "FOSCiting", "CitedId", "FOSCited"]

PaperReferencesFields.to_csv("./Processed_data/MAG_ALLReferenceL0FieldMerged.tsv", index=None, sep="\t")

PaperReferencesFields
