==> 02_NMI_TEST.ipynb <==
import pandas as pd
from matplotlib import pyplot as plt
import glob

datadir = "./MAG_L2_Similarity/"
l2_nmi_files = glob.glob(datadir + "*.tsv")
print(l2_nmi_files)
df_l2_nmi = pd.concat((pd.read_csv(f, sep="\t") for f in l2_nmi_files))

def lst_elem_to_float(lst):
    return [float(x) for x in lst]

def cm2inch(value):
    return value/2.54

df_l2_nmi["LayerRatio"] = df_l2_nmi["LayerRatio"].map(lambda x: x.strip("[").strip("]").split(",")).map(lst_elem_to_float).map(tuple)

df_l2_nmi

DF_FOS_LIST = df_l2_nmi["L0FOS"].unique()

df_l2_nmi["NetworkType"].unique()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.5, 0.5)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].plot(label = "BC-CC-Hybrid (0.5, 0.5)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.1, 0.9)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].plot(label = "BC-CC-Hybrid (CC0.1, BC0.9)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.9, 0.1)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].plot(label = "BC-CC-Hybrid (CC0.9, BC0.1)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-CC")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 1, 0)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].plot(label = "BC-CC-Hybrid (CC1, BC0)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 0, 1)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].plot(label = "BC-CC-Hybrid (CC0, BC1)", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.savefig("ExpFigures/20201001_BC_CC_Hybrid_NMI_GRAN.png")
plt.show()
plt.close()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.5, 0.5)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].plot(label = "BC-CC-Hybrid (0.5, 0.5)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.1, 0.9)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].plot(label = "BC-CC-Hybrid (CC0.1, BC0.9)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.9, 0.1)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].plot(label = "BC-CC-Hybrid (CC0.9, BC0.1)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-CC")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 1, 0)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].plot(label = "BC-CC-Hybrid (CC1, BC0)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 0, 1)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].plot(label = "BC-CC-Hybrid (CC0, BC1)", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.savefig("ExpFigures/20201001_BC_CC_Hybrid_NMI_RES.png")
plt.show()
plt.close()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.5, 0.5)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].plot(label = "BC-CC-Hybrid (0.5, 0.5)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.1, 0.9)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].plot(label = "BC-CC-Hybrid (CC0.1, BC0.9)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.9, 0.1)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].plot(label = "BC-CC-Hybrid (CC0.9, BC0.1)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/3, CC1/3, BC1/3)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/10, 9/20, 9/20)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/10, CC9/20, BC9/20)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T2")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid-T2(DC1/3, CC1/3, BC1/3)", ax = now_ax)    
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")
plt.tight_layout()
#plt.savefig("ExpFigures/20201001_BC_CC_DC_Hybrid_NMI_GRAN.png")
plt.show()
plt.close()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.5, 0.5)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].plot(label = "BC-CC-Hybrid (0.5, 0.5)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.1, 0.9)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].plot(label = "BC-CC-Hybrid (CC0.1, BC0.9)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.9, 0.1)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].plot(label = "BC-CC-Hybrid (CC0.9, BC0.1)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/3, CC1/3, BC1/3)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/10, 9/20, 9/20)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/10, CC9/20, BC9/20)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T2")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid-T2(DC1/3, CC1/3, BC1/3)", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")
plt.tight_layout()
plt.savefig("ExpFigures/20201001_BC_CC_DC_Hybrid_NMI_RES.png")
plt.show()
plt.close()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "DC-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1, 0, 0)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].plot(label = "DC-T1", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "DC-T2")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 0, 0, 1)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].plot(label = "DC-T2", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/3, CC1/3, BC1/3)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/10, 9/20, 9/20)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/10, CC9/20, BC9/20)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T2")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("Granularity")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid-T2(DC1/3, CC1/3, BC1/3)", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")   
plt.tight_layout()
plt.savefig("ExpFigures/20201001_DC_Hybrid_NMI_GRAN.png")
plt.show()
plt.close()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "DC-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1, 0, 0)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].plot(label = "DC-T1", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "DC-T2")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 0, 0, 1)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].plot(label = "DC-T2", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/3, CC1/3, BC1/3)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/10, 9/20, 9/20)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/10, CC9/20, BC9/20)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T2")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["L2NMI"].sort_values().plot(label = "BC-CC-DC-Hybrid-T2(DC1/3, CC1/3, BC1/3)", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")   
plt.tight_layout()
plt.savefig("ExpFigures/20201001_DC_Hybrid_NMI_RES.png")
plt.show()
plt.close()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.5, 0.5)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].plot(label = "BC-CC-Hybrid (0.5, 0.5)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.1, 0.9)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].plot(label = "BC-CC-Hybrid (CC0.1, BC0.9)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-Hybrid")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0,0.9, 0.1)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].plot(label = "BC-CC-Hybrid (CC0.9, BC0.1)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/3, CC1/3, BC1/3)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/10, 9/20, 9/20)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/10, CC9/20, BC9/20)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T2")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].sort_values().plot(label = "BC-CC-DC-Hybrid-T2(DC1/3, CC1/3, BC1/3)", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("Granularity")
plt.tight_layout()
plt.savefig("ExpFigures/20201001_BC_CC_DC_Hybrid_GRAN_RES.png")
plt.show()
plt.close()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "DC-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1, 0, 0)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].plot(label = "DC-T1", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "DC-T2")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 0, 0, 1)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].plot(label = "DC-T2", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/3, CC1/3, BC1/3)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T1")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(1/10, 9/20, 9/20)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].sort_values().plot(label = "BC-CC-DC-Hybrid (DC1/10, CC9/20, BC9/20)", ax = now_ax)
    df_l2_nmi[(df_l2_nmi["NetworkType"] == "NS-BC-CC-DC-Hybrid-T2")
              & (df_l2_nmi["LayerRatio"].map(tuple).isin([(0, 1/3, 1/3, 1/3)]))
              & (df_l2_nmi["L0FOS"] == FOS)].set_index("ResParm")["Granularity"].sort_values().plot(label = "BC-CC-DC-Hybrid-T2(DC1/3, CC1/3, BC1/3)", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("Granularity")
plt.tight_layout()
plt.savefig("ExpFigures/20201001_DC_GRAN_RES.png")
plt.show()
plt.close()

df_l2_nmi[(df_l2_nmi["NetworkType"] == "DC-T1")]
