==> 03_Preprocess_GetCitationCounts.ipynb <==
import csv
from collections import defaultdict
from tqdm.notebook import tqdm
import pandas as pd

input_file_reference = "./MAG_dataset/mag/PaperReferences.txt"
citation_number_dict_compare = defaultdict(int)
test_duplicated_link2 = defaultdict(int)
duplicates2 = 0
with open(input_file_reference, 'r') as csvfile:
    reader = csv.reader(csvfile, delimiter='\t')
    next(reader)   
    for row in tqdm(reader): # for a new file
    #for row in tqdm(reader, total=956198971): # Hard coded to improve the speed
        if(test_duplicated_link2[(int(row[0]), int(row[1]))] != 0):
            duplicates2 += 1
        test_duplicated_link2[(int(row[0]), int(row[1]))] += 1     
        citation_number_dict_compare[int(row[1])] += 1

duplicates2

df_citation_num_compare = pd.Series(citation_number_dict_compare).reset_index()
df_citation_num_compare.columns = ["PaperId", "CitationCountsCompare"]

input_file_reference_with_l0 = "./fastnvme01//MAG_ALLReferenceL0FieldMerged.tsv"
field_edge_list_dict = defaultdict(list)
citation_number_dict = defaultdict(int)
citation_number_dict_main_field = defaultdict(lambda: defaultdict(int))
with open(input_file_reference_with_l0, 'r') as csvfile:
    reader = csv.reader(csvfile, delimiter='\t')
    next(reader)
    #for row in tqdm(reader, total=get_filelength(input_file_reference_with_l0)): # for a new file
    for row in tqdm(reader, total=956198971): # Hard coded to improve the speed        
        if(row[1] == row[3]):
            citation_number_dict_main_field[int(row[1])][int(row[2])] += 1

duplicates

df_citation_num = pd.Series(citation_number_dict_compare).reset_index()
df_citation_num.columns = ["PaperId", "CitationCounts"]

df_citation_num.to_csv("./Processed_data/MAG_CitationCounts.tsv", index=None, sep="\t")

list_field_citation = []
for field, secdict in citation_number_dict_main_field.items():
    for paperid, value in secdict.items():
        list_field_citation.append((paperid, field, value))
        #print(key, key2, value)

df_citation_num_main_field = pd.DataFrame(list_field_citation)
df_citation_num_main_field.columns = ["PaperId", "L0Field", "InFieldCitationCounts"]

df_citation_num_main_field.to_csv("./Processed_data/MAG_L0CitationCounts.tsv", sep="\t", index=None)
