==> 04_NMI_TEST_BY_METADATA_INFIELD_QUARTILE.ipynb <==
import pandas as pd
from matplotlib import pyplot as plt
import glob
from collections import defaultdict
from collections import Counter
import csv
from clusim.clustering import Clustering, print_clustering
import clusim.sim as sim
import math
from multiprocessing import Pool
from tqdm.notebook import tqdm
import itertools

def lst_elem_to_float(lst):
    return [float(x) for x in lst]

def cm2inch(value):
    return value/2.54

quantile_df = pd.read_csv("./MAG_L2_Similarity/20201028_NMI_BY_CITATION_INFILED_QUANTILE_RE.tsv", sep="\t")

quantile_df["net_type"].unique()

def stup_to_ftup(tup):
    tup = eval(tup)
    reslist = [float(x) for x in tup]
    return tuple(reslist)

DF_FOS_LIST = quantile_df["field_name"].unique()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    #now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T1 LQ1", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T2 LQ1", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("Granularity")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    #now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T1 LQ2", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T2 LQ2", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("Granularity")
plt.tight_layout()    
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    #now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T1 LQ3", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T2 LQ3", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("Granularity")
plt.tight_layout()    
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    #now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T1 LQ4", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T2 LQ4", ax = now_ax)
    now_ax.legend()
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")
plt.tight_layout()
plt.legend()
plt.show()

blue_colors = [(0, 0.3, 0.4), (0, 0.3, 0.6), (0, 0.3, 0.8), (0, 0.3, 1)]
red_colors = [(0.4, 0, 0), (0.6, 0, 0), (0.8, 0, 0), (1.0, 0, 0)]

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    #now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T1 LQ1", ax = now_ax, c=blue_colors[0])
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T1 LQ2", ax = now_ax, c=blue_colors[1])
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T1 LQ3", ax = now_ax, c=blue_colors[2])
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T1 LQ4", ax = now_ax, c=blue_colors[3])
    
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T2 LQ1", ax = now_ax, c=red_colors[0])
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T2 LQ2", ax = now_ax, c=red_colors[1])
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T2 LQ3", ax = now_ax, c=red_colors[2])
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("res").sort_index()["nmi"].plot(label = "DC-T2 LQ4", ax = now_ax, c=red_colors[3])
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

blue_colors = [(0, 0.3, 0.4), (0, 0.3, 0.6), (0, 0.3, 0.8), (0, 0.3, 1)]
red_colors = [(0.4, 0, 0), (0.6, 0, 0), (0.8, 0, 0), (1.0, 0, 0)]

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T1 LQ1", ax = now_ax, c=blue_colors[0])
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T1 LQ2", ax = now_ax, c=blue_colors[1])
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T1 LQ3", ax = now_ax, c=blue_colors[2])
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T1 LQ4", ax = now_ax, c=blue_colors[3])
    
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ1", ax = now_ax, c=red_colors[0])
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ2", ax = now_ax, c=red_colors[1])
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ3", ax = now_ax, c=red_colors[2])
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ4", ax = now_ax, c=red_colors[3])
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

blue_colors = [(0, 0.3, 0.4), (0, 0.3, 0.6), (0, 0.3, 0.8), (0, 0.3, 1)]
red_colors = [(0.4, 0, 0), (0.6, 0, 0), (0.8, 0, 0), (1.0, 0, 0)]

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T1 LQ1", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ1", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T1 LQ2", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ2", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T1 LQ3", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ3", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "DC-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(1, 0, 0)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T1 LQ4", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ4", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

blue_colors = [(0, 0.3, 0.4), (0, 0.3, 0.6), (0, 0.3, 0.8), (0, 0.3, 1)]
red_colors = [(0.4, 0, 0), (0.6, 0, 0), (0.8, 0, 0), (1.0, 0, 0)]

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid DC-T1 LQ1", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid DC-T2 LQ1", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid DC-T1 LQ2", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid DC-T2 LQ2", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid DC-T1 LQ3", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid DC-T2 LQ3", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T1")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid DC-T1 LQ4", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid DC-T2 LQ4", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

blue_colors = [(0, 0.3, 0.4), (0, 0.3, 0.6), (0, 0.3, 0.8), (0, 0.3, 1)]
red_colors = [(0.4, 0, 0), (0.6, 0, 0), (0.8, 0, 0), (1.0, 0, 0)]

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-Hybrid")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.9, 0.1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC LQ1", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC-DC-T2 LQ1", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-Hybrid")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.9, 0.1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC LQ1", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC-DC-T2 LQ1", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-Hybrid")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.9, 0.1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC LQ3", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC-DC-T2 LQ3", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-Hybrid")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.9, 0.1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC LQ4", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC-DC-T2 LQ4", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

blue_colors = [(0, 0.3, 0.4), (0, 0.3, 0.6), (0, 0.3, 0.8), (0, 0.3, 1)]
red_colors = [(0.4, 0, 0), (0.6, 0, 0), (0.8, 0, 0), (1.0, 0, 0)]

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC-DC-T2 LQ1", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ1", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC-DC-T2 LQ2", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.25) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ2", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC-DC-T2 LQ3", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.5) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ3", ax = now_ax)
    now_ax.legend()    
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()

fig, axes = plt.subplots(2, 3, figsize =(cm2inch(19)*2, cm2inch(11)*2))
axes = axes.flatten()
for idx, FOS in enumerate(DF_FOS_LIST):
    now_ax = axes[idx]    
    now_ax.set_xscale("log")
    now_ax.set_yscale("log")
    quantile_df[(quantile_df["net_type"] == "NS-BC-CC-DC-Hybrid-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0.33, 0.33, 0.33)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("granularity").sort_index()["nmi"].plot(label = "Hybrid BC-CC-DC-T2 LQ4", ax = now_ax)
    quantile_df[(quantile_df["net_type"] == "DC-T2")
              & (quantile_df["r_layer"].map(stup_to_ftup).isin([(0, 0, 0, 1)]))
              & (quantile_df["field_name"] == FOS)
              & (quantile_df["pct_lb"] ==0.75) ].set_index("granularity").sort_index()["nmi"].plot(label = "DC-T2 LQ4", ax = now_ax)
    now_ax.set_title(FOS)
    now_ax.set_ylabel("NMI")    
plt.tight_layout()
plt.legend()
plt.show()
